Convert to Kotlin JUnit
