+++ build.gradle.kts
@@ -1,10 +1,9 @@
+import org.gradle.api.tasks.testing.logging.TestExceptionFormat.FULL
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
     id("org.springframework.boot") version "2.4.0-M4"
     id("io.spring.dependency-management") version "1.0.10.RELEASE"
-    id("java")
-    id("groovy")
     id("org.sonarqube") version "3.0"
     id("org.jetbrains.kotlin.jvm") version "1.4.10"
 }
@@ -28,17 +27,13 @@ dependencies {
         exclude("org.junit.vintage", "junit-vintage-engine")
     }
     implementation("org.jsoup:jsoup:1.13.1")
-    compileOnly("org.projectlombok:lombok")
-    annotationProcessor("org.projectlombok:lombok")
     implementation("org.apache.commons:commons-lang3")
-    testImplementation("org.junit.jupiter:junit-jupiter-api")
-    testImplementation("org.junit.vintage:junit-vintage-engine")
-    testImplementation("org.spockframework:spock-core:2.0-M3-groovy-3.0")
-    implementation("org.codehaus.groovy:groovy:3.0.5")
-    implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
 }
 tasks.withType<Test> {
     useJUnitPlatform()
+    testLogging {
+        exceptionFormat = FULL
+    }
 }
 tasks.withType<KotlinCompile> {
     kotlinOptions {
