--- src/main/java/com/webpage/checker/Checker.kt
@@ -1,9 +1,0 @@
-package com.webpage.checker
-
-internal class Checker {
-    companion object {
-        const val AUDIOBOOKS = "https://bridges.overdrive.com/bridges-kirkendall/content/collection/37479"
-        const val EBOOKS = "https://bridges.overdrive.com/bridges-kirkendall/content/collection/37473"
-        const val SECTION_LIST_CLASS = "Facets-sectionListItem  "
-    }
-}
